+++ lang-haskell/app-view/comp-view/FlareTiming/Penal/EssGoal.hs
@@ -280,7 +280,7 @@ pointRow w _earliest cTime cArrival dfNt pt tp sEx x = do
                     Nothing -> ("", "", "")
                     Just
                         Alt.AltBreakdown
-                            { place = nth
+                            { placeTaken = nth
                             , total = p@(TaskPoints pts)
                             } -> (showRank nth, showRounded pts, showTaskPointsDiff p p'))